pyproject.toml
[project]
name = "system-documentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocs-git-revision-date-localized-plugin>=1.5.0",
    "mkdocs-macros-plugin>=1.4.1",
    "mkdocs-material>=9.7.0",
    "mkdocs-swagger-ui-tag>=0.7.2",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.1",
    "pytest-json-report>=1.5.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ruff>=0.14.4",
]
